Refactor Landing Page Components Names
