+++ client/src/App.tsx
@@ -3,13 +3,13 @@ import Servers from "./pages/Servers";
 import ChannelsSidebar from "./components/ChannelsSidebar";
 import Chat from "./components/Chat";
 import WelcomeToServer from "./components/WelcomeToServer";
-import Home from "./pages/Home";
+import Landing from "./pages/Landing";
 
 function App() {
   return (
     <>
       <Routes>
-        <Route path="/" element={<Home />} />
+        <Route path="/" element={<Landing />} />
         <Route path="/channels/" element={<Servers />}>
           <Route path=":serverId" element={<ChannelsSidebar />}>
             <Route index element={<WelcomeToServer />} />
